Separate main program from library calls
