+++ pynest/pynest.py
@@ -1,0 +1,136 @@
+#! /usr/bin/python
+
+# nest.py -- a python interface to the Nest Thermostat
+# by Scott M Baker, smbaker@gmail.com, http://www.smbaker.com/
+#
+# Usage:
+#    'nest.py help' will tell you what to do and how to do it
+#
+# Licensing:
+#    This is distributed unider the Creative Commons 3.0 Non-commecrial,
+#    Attribution, Share-Alike license. You can use the code for noncommercial
+#    purposes. You may NOT sell it. If you do use it, then you must make an
+#    attribution to me (i.e. Include my name and thank me for the hours I spent
+#    on this)
+#
+# Acknowledgements:
+#    Chris Burris's Siri Nest Proxy was very helpful to learn the nest's
+#       authentication and some bits of the protocol.
+
+import urllib
+import urllib2
+
+try:
+   import json
+except ImportError:
+   import simplejson as json
+   
+class Nest:
+    def __init__(self, username, password, serial=None, index=0, units="F"):
+        self.username = username
+        self.password = password
+        self.serial = serial
+        self.units = units
+        self.index = index
+
+    def loads(self, res):
+        if hasattr(json, "loads"):
+            res = json.loads(res)
+        else:
+            res = json.read(res)
+        return res
+
+    def login(self):
+        data = urllib.urlencode({"username": self.username, "password": self.password})
+
+        req = urllib2.Request("https://home.nest.com/user/login",
+                              data,
+                              {"user-agent":"Nest/1.1.0.10 CFNetwork/548.0.4"})
+
+        res = urllib2.urlopen(req).read()
+
+        res = self.loads(res)
+
+        self.transport_url = res["urls"]["transport_url"]
+        self.access_token = res["access_token"]
+        self.userid = res["userid"]
+
+    def get_status(self):
+        req = urllib2.Request(self.transport_url + "/v2/mobile/user." + self.userid,
+                              headers={"user-agent":"Nest/1.1.0.10 CFNetwork/548.0.4",
+                                       "Authorization":"Basic " + self.access_token,
+                                       "X-nl-user-id": self.userid,
+                                       "X-nl-protocol-version": "1"})
+
+        res = urllib2.urlopen(req).read()
+
+        res = self.loads(res)
+
+        self.structure_id = res["structure"].keys()[0]
+
+        if (self.serial is None):
+            self.device_id = res["structure"][self.structure_id]["devices"][self.index]
+            self.serial = self.device_id.split(".")[1]
+
+        self.status = res
+
+        #print "res.keys", res.keys()
+        #print "res[structure][structure_id].keys", res["structure"][self.structure_id].keys()
+        #print "res[device].keys", res["device"].keys()
+        #print "res[device][serial].keys", res["device"][self.serial].keys()
+        #print "res[shared][serial].keys", res["shared"][self.serial].keys()
+
+    def temp_in(self, temp):
+        if (self.units == "F"):
+            return (temp - 32.0) / 1.8
+        else:
+            return temp
+
+    def temp_out(self, temp):
+        if (self.units == "F"):
+            return temp*1.8 + 32.0
+        else:
+            return temp
+
+    def show_status(self):
+        shared = self.status["shared"][self.serial]
+        device = self.status["device"][self.serial]
+
+        allvars = shared
+        allvars.update(device)
+
+        for k in sorted(allvars.keys()):
+             print k + "."*(32-len(k)) + ":", allvars[k]
+
+    def show_curtemp(self):
+        temp = self.status["shared"][self.serial]["current_temperature"]
+        temp = self.temp_out(temp)
+
+        print "%0.1f" % temp
+
+    def set_temperature(self, temp):
+        temp = self.temp_in(temp)
+
+        data = '{"target_change_pending":true,"target_temperature":' + '%0.1f' % temp + '}'
+        req = urllib2.Request(self.transport_url + "/v2/put/shared." + self.serial,
+                              data,
+                              {"user-agent":"Nest/1.1.0.10 CFNetwork/548.0.4",
+                               "Authorization":"Basic " + self.access_token,
+                               "X-nl-protocol-version": "1"})
+
+        res = urllib2.urlopen(req).read()
+
+        print res
+
+    def set_fan(self, state):
+        data = '{"fan_mode":"' + str(state) + '"}'
+        req = urllib2.Request(self.transport_url + "/v2/put/device." + self.serial,
+                              data,
+                              {"user-agent":"Nest/1.1.0.10 CFNetwork/548.0.4",
+                               "Authorization":"Basic " + self.access_token,
+                               "X-nl-protocol-version": "1"})
+
+        res = urllib2.urlopen(req).read()
+
+        print res
+
