+++ setup.py
@@ -9,4 +9,5 @@ setup(name='pynest',
       author_email='smbaker@gmail.com',
       url='http://www.smbaker.com/',
       scripts=['nest.py'],
+      packages=['pynest'],
      )
